Split TS configs for typechecking + building; run parity tests in CI
